ask xcrun(1) for the sourcekit-lsp server path
Recent XCode ships with the `sourcekit-lsp` server, but it is a "developer tool"
and so not on the normal path.  

This updates the logic for discovering the executable to:
1. use a `lsp-dependency` to specify the executable, and
2. try `xcrun --find` if the executable is not on the path.

This makes the platform-supplied server work for me, and
is generally how apple treat these development tools.

Additionally, the server is registered with `lsp-mode`
automatically, through an autoload, so the package will
just work™ after installation.